feat: Add support for Raspberry Pi 4B Bookworm OS
This commit adds support for the Raspberry Pi 4B Bookworm OS by including the necessary pinout configuration file and controller type definition. The pinout file 'raspi-4b-bookworm.json' is added to the 'pinouts' directory, and a new controller type entry is added to the 'valueMaps' class in 'Constants.ts'. This allows the application to recognize and work with the Raspberry Pi 4B Bookworm OS as a valid controller type.